Rename the bazeltest package to app, and othermodule to module. Restrict the visibility of tests to private.
