+++ BUILD
@@ -1,7 +1,7 @@
 scala_binary(
     name = "App",
+    main_class = "app.Main",
     deps = [
-        "//bazeltest"
+        "//app",
     ],
-    main_class = "bazeltest.Main"
 )
