+++ bazel-scala_play
@@ -1,0 +1,1 @@
+/home/l-8296/.cache/bazel/_bazel_l-8296/76d33ba2f8986abe556e65dc46b40a5c/execroot/__main__
