+++ module/src/main/scala/module/Worker.scala
@@ -1,4 +1,4 @@
-package othermodule
+package module
 
 class Worker {
